initial course function migration to settings
