+++ backend/functions/src/course/functions.ts
@@ -1,32 +1,8 @@
 import { db } from '../config'
 import { getStudentsData } from '../student/functions'
 import admin from 'firebase-admin'
-import { Course, FirestoreCourse, FirestoreGroup, Semester } from '../types'
+import { Course, FirestoreCourse, FirestoreGroup } from '../types'
 const courseRef = db.collection('courses')
-const semesterRef = db.collection('utils').doc('semester')
-
-export const getCurrentSemester = async (): Promise<String> => {
-  const semData = (await semesterRef.get()).data() as Semester
-  return semData.currentSemester
-}
-
-export const setCurrentSemester = async (sem: String) => {
-  return semesterRef.set({ currentSemester: sem, surveyOpen: false })
-}
-
-export const getAllSemesters = async () => {
-  const semData = (await semesterRef.get()).data() as Semester
-  return [semData.currentSemester, ...semData.allSemesters]
-}
-
-export const getSurveyStatus = async () => {
-  const semData = (await semesterRef.get()).data() as Semester
-  return semData.surveyOpen
-}
-
-export const setSurveyStatus = async (status: Boolean) => {
-  return semesterRef.update({ surveyOpen: status })
-}
 
 async function getCourseInfo(courseId: string) {
   const snapshot = await courseRef.doc(courseId).get()
